# Improve speed of sample spreading - much faster inference